+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 * removed `SerializedLog`. Use instead `EntityFileLog\Log\Engine\EntityFileLog`.
 
 ## 2.25 branch
+### 2.25.2
+* `AuthComponent` now uses the `auth` find method (`UsersTable::findAuth()`);
+* fixed bug on required data check for mailer classes.
+
 ### 2.25.1
 * uses the `cakephp-entity-file-log` package, so now `SerializedLog` extends
     `EntityFileLog\Log\Engine\EntityFileLog`. In a future release